+++ metrics/metrics.go
@@ -85,3 +85,18 @@ func SendTimeout(ctx context.Context, d time.Duration) {
 	newrelic.SendTimeout(ctx, d)
 	datadog.SendTimeout(ctx, d)
 }
+
+func ObserveBufferSize(t string, size int) {
+	prometheus.ObserveBufferSize(t, size)
+	datadog.ObserveBufferSize(t, size)
+}
+
+func SetBufferDefaultSize(t string, size int) {
+	prometheus.SetBufferDefaultSize(t, size)
+	datadog.SetBufferDefaultSize(t, size)
+}
+
+func SetBufferMaxSize(t string, size int) {
+	prometheus.SetBufferMaxSize(t, size)
+	datadog.SetBufferMaxSize(t, size)
+}