+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Add
 - Add support of 16-bit BMP.
 - Add `IMGPROXY_NEW_RELIC_LABELS` config.
+- Add sending additional metrics to Datadog and `IMGPROXY_DATADOG_ENABLE_ADDITIONAL_METRICS` config.
 
 ### Fix
 - Fix trimming of CMYK images.